[REFACTOR] 기존 Guest 관련 클래스를 의존하는 테스트 클래스 import 문 수정 (#153)
