+++ src/test/java/org/chzzk/howmeet/domain/temporary/auth/service/GuestSaveServiceTest.java
@@ -1,8 +1,9 @@
 package org.chzzk.howmeet.domain.temporary.auth.service;
 
 import org.chzzk.howmeet.domain.common.model.EncodedPassword;
-import org.chzzk.howmeet.domain.temporary.auth.entity.Guest;
-import org.chzzk.howmeet.domain.temporary.auth.repository.GuestRepository;
+import org.chzzk.howmeet.domain.temporary.guest.entity.Guest;
+import org.chzzk.howmeet.domain.temporary.guest.repository.GuestRepository;
+import org.chzzk.howmeet.domain.temporary.guest.service.GuestSaveService;
 import org.chzzk.howmeet.domain.temporary.schedule.exception.GSException;
 import org.chzzk.howmeet.domain.temporary.schedule.repository.GSRepository;
 import org.junit.jupiter.api.DisplayName;
